BIG CHANGE: detail views are now open in modals. I made that change because IONIC and UIROUTER are not smart enough to allow you to show in your current view a shared template/controller. Which means you have to duplicate state configuration in order to reuse the same component (template/controller). This not good enough. So I change my design and I now open detail views in a modal like Google do for all its apps. This is a shame I had to sacrifice the swipe back functionality on iOS. Maybe IONIC will add a swipe back behavior on modal later.
In order to do that I had to take care of the back button myself (so it changes upon platforms) and I had to create my own image cache directive because they are using an isolate scope without reason which disable our ability to use it on elements already binded to an isolated scope. I also had to create a custom animation for modals (only up is built in).

I fixed a bug when the description contains html tags like ul/li which doesn't work with the css effect I apply on texts in detail views. A simple filter made the trick.

I made some smart refactoring.

It is now possible to show the detail view in the favourites screen when clicking on a bookmarked item.